bIO/49, Make FastaFlow not exect lines anymore.
- As of now, FastaFlow expects its input to be lines. This it not
  garanteed at the type level so it is doomed to be broken.
- You can now only create FastaFlows that have a framing flow prepended.
- Small cleanups.
