+++ stream/src/test/resources/at/bioinform/codec/fasta/fasta_easy.fa
