+++ stream/src/test/resources/at/bioinform/codec/fasta/fasta_with_descriptions.fa
