+++ stream/src/test/resources/at/bioinform/codec/lucene/fasta_easy.fa
